+++ ancient-src/bsd211/h/stat.h
@@ -13,20 +13,20 @@ struct	stat
 {
 	dev_t	st_dev;
 	ino_t	st_ino;
-	unsigned short st_mode;
-	short	st_nlink;
+	uint16_t st_mode;
+	int16_t	st_nlink;
 	uid_t	st_uid;
 	gid_t	st_gid;
 	dev_t	st_rdev;
 	off_t	st_size;
 	time_t	st_atime;
-	int	st_spare1;
+	int16_t	st_spare1;
 	time_t	st_mtime;
-	int	st_spare2;
+	int16_t	st_spare2;
 	time_t	st_ctime;
-	int	st_spare3;
-	long	st_blksize;
-	long	st_blocks;
+	int16_t	st_spare3;
+	int32_t	st_blksize;
+	int32_t	st_blocks;
 	u_short	st_flags;
 	u_short	st_spare4[3];
 };
